services/fetcher/requirements.txt
# Add service-specific dependencies here
requests
confluent-kafka
python-dotenv
